correct prototype

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: hramaros <hramaros@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/19 15:46:44 by hramaros          #+#    #+#             */
-/*   Updated: 2024/02/21 13:50:35 by hramaros         ###   ########.fr       */
+/*   Updated: 2024/02/22 13:42:52 by hramaros         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	int	i;
