+++ ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: hramaros <hramaros@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/19 15:13:35 by hramaros          #+#    #+#             */
-/*   Updated: 2024/02/21 13:05:56 by hramaros         ###   ########.fr       */
+/*   Updated: 2024/02/22 13:50:11 by hramaros         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	void	*s_cpy;
